# docs/api.yaml
openapi: 3.0.0
info:
  title: HRM API
  description: API documentation for the Human Resource Management system.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api # Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...
paths:
  /admin/login:
    post:
      summary: Admin login
      operationId: adminLogin
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: admin@gmail.com
                password:
                  type: string
                  example: 1
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  data:
                    type: object
                    properties:
                      adminId:
                        type: integer
                        example: 1
                      email:
                        type: string
                        example: admin@gmail.com
                      role:
                        type: string
                        example: ADMIN
        '400':
          description: Bad Request - Missing email or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Please provide email and password
        '401':
          description: Unauthorized - Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Invalid credentials or not an admin
...
  /onboarding:
    post:
      tags:
        - Recruitment - Onboarding
      summary: Create an Onboarding Record
      operationId: createOnboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                departmentId:
                  type: integer
                  description: ID of the employee.
                jobApplicationId:
                  type: integer
                  description: ID of the job application.
                designationId:
                  type: integer
                  description: ID of the designation.
              required:
                - employeeId
                - jobApplicationId
                - designationId
      responses:
        '201':
          description: Onboarding record created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      onboarding:
                        type: object
                        description: The created onboarding record.
        '400':
          description: Bad request if required fields are missing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string

    get:
      tags:
        - Recruitment - Onboarding
      summary: Get All Onboarding Records
      operationId: getAllOnboardings
      responses:
        '200':
          description: List of onboarding records retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  results:
                    type: integer
                    description: Number of onboarding records found.
                  data:
                    type: object
                    properties:
                      onboardings:
                        type: array
                        items:
                          type: object
                          description: An onboarding record.
        '404':
          description: No onboarding records found.
  /onboarding/{id}:
    get:
      tags:
        - Recruitment - Onboarding
      summary: Get an Onboarding Record by ID
      operationId: getOnboardingById
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the onboarding record.
          schema:
            type: integer
      responses:
        '200':
          description: Onboarding record retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      onboarding:
                        type: object
                        description: The requested onboarding record.
        '404':
          description: No onboarding record found with that ID.

    patch:
      tags:
        - Recruitment - Onboarding
      summary: Update an Onboarding Record
      operationId: updateOnboarding
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the onboarding record.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                designationId:
                  type: integer
                  description: Updated designation ID.
                status:
                  type: string
                  description: Updated status.
                completionDate:
                  type: string
                  format: date
                  description: Completion date for onboarding.
      responses:
        '200':
          description: Onboarding record updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      onboarding:
                        type: object
                        description: The updated onboarding record.
        '404':
          description: No onboarding record found with that ID.

    delete:
      tags:
        - Recruitment - Onboarding
      summary: Delete an Onboarding Record
      operationId: deleteOnboarding
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the onboarding record.
          schema:
            type: integer
      responses:
        '204':
          description: Onboarding record deleted successfully.
        '404':
          description: No onboarding record found with that ID.
  /notification:
    post:
      tags:
        - Notifications & Alerts - System Announcement
      summary: Create a System Announcement
      operationId: createSystemAnnouncement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: The title of the announcement.
                message:
                  type: string
                  description: The content of the announcement message.
                targetGroup:
                  type: string
                  description: The group targeted for the announcement (e.g., "ALL_EMPLOYEES", "HR", "DEPARTMENT").
                departmentId:
                  type: integer
                  description: Department ID for department-specific announcements.
                validUntil:
                  type: string
                  format: date
                  description: The expiration date of the announcement.
              required:
                - title
                - message
                - targetGroup
      responses:
        '201':
          description: System announcement created and notifications sent to specified recipients.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      announcement:
                        type: object
                        description: The created announcement details.
                      employeesNotified:
                        type: integer
                        description: The number of employees notified.
        '400':
          description: Missing required fields or invalid target group.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Title, message, and target group are required.'
        '404':
          description: No employees found for specified departmentId.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'No employees found for department ID.'
...